Fix scpaste-index remote path concatenation
Hi, I found what I believe to be is a bug in scpaste as I couldn't
generate and index for my pastes. This is my first time sending a patch
using email so hopefully I'm typing this in the right area :-)

/getz

ls $path on the remote machine wouldn't execute as it requires being wrapped
in quotes.